Retry creating the search index when it fails
When initially creating a cluster from a database, each registry
competes for creating the search index. Only the first one to complete
succeeds since it adds all of the entries at the same time.

If the setup fails due to a database constraint, wait a random amount of
time (1-10 seconds) and retry creating the search index.